# Clone child elements lazily in Vec