@aswaterman 2020-02-11
Thanks again for prototyping this idea! It has been a while since we saw that magnitude of perf improvement.