@davidbiancolin 2020-02-11
See, Bordeaux does make good wine.