@aswaterman 2020-02-11
lol

On Tue, Feb 11, 2020 at 11:18 AM David Biancolin <notifications@github.com>
wrote:

> See, Bordeaux does make good wine.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/freechipsproject/chisel3/pull/1329?email_source=notifications&email_token=AAH3XQSEP4A2GZDG5XGRPYTRCLMWVA5CNFSM4KS2R6X2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOELNITGY#issuecomment-584747419>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAH3XQWUZTNL7JM2OOYOV6DRCLMWVANCNFSM4KS2R6XQ>
> .
>